LcpFinder Class

This S4 class is a wrapper around a CppLcpFinder C++ object that is made available to R via the 'Rcpp' package. Instances of this class can be created from a Quadtree object using the lcp_finder function.

The methods of the C++ object (CppLcpFinder) can be accessed from R, but the typical end-user should have no need of these methods - they are meant for internal use. That being said, descriptions of the available methods can be found on the CppLcpFinder documentation page.

Details

Functions for creating a LcpFinder object:

lcp_finder()

Methods:

find_lcp()

find_lcps()

plot()

show()

summarize_lcps()

summary()

Slots

ptr

a C++ object of class CppLcpFinder